Refactor/trustees (#559)
* refactor bridge-common bridge-features

* transition_trustee_session

* session info providor

* half of rpc

* refactor rpc finish

* fix genesis for refactor

* crosschain binding init

* crosschain binding finish

* final

* fix bugs

* support for new events and trustees check for call

+++ xrml/xmultisig/src/types.rs
@@ -34,6 +34,6 @@ pub struct AddrInfo<AccountId> {
 #[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
 pub struct PendingState<Proposal> {
     pub yet_needed: u32,
-    pub owners_done: u32,
+    pub owners_done: u64,
     pub proposal: Box<Proposal>,
 }
